Deleting Mtrie entries asynchronously
The deletion of route entry in Vrouter can lead to a deletion of
complete mtrie bucket. Right now this bucket deletion is executed after
a synchronus RCU mechanism. All of this happens in the context of Agent
thread leading to slow route deletion process.

To make the deletion faster, asynchronous version of RCU mechansism is
used. With this, the nexthop deletion would also be invoked from
different CPU's, bringing a contention for its ref_cnt. So the nexthops
ref_cnt drecrement and its comparison to zero is made atomic.

While doing the 'rmmod' of vrouter, we execute rcu_barrier() to ensure
that all the rcu callbacks are invoked. Post this rcu barrier, no
further asynchronous rcu synchronisation be invoked and this is
controlled by making vr_not_ready to 1 for module exit also.

closes-bug: #1465195
Change-Id: I53f087a049c37a83fcee4bb312f05609c5513632