Runtime parameters in the vrouter utility.
vRouter kernel module exports some of runtime parameters using
the sysctl interface. Allow to set those parameters using the
vrouter utility as well.

This change was introduced to allow vRouter/DPDK to be configurable
in similar fashion the kernel module is. As vRouter/DPDK is an userspace
application, therefore it cannot use the sysctl interface. From now on,
the vrouter utility may be used for the same purpose as sysctl.

No additional changes were required to allow the kernel module to be
configured with the vrouter utility as well, so it may benefit from
this additional feature. The change does not interfere with the sysctl
method, they can be used alternatively.

The vrouter utility commandline parameters have the same names as files
in /proc/sys/net/vrouter:
   perfr, perfs, from_vm_mss_adj, to_vm_mss_adj, perfr1, perfr2, perfr3,
   perfp, perfq1, perfq2, perfq3, udp_coff, flow_hold_limit, mudp.
In the kernel module version all of them can be configured using the
vrouter utility, as well as the sysctl interface. Currently, vRouter/DPDK
supports only three of those features: from_vm_mss_adj, flow_hold_limit,
mudp. The state of every feature is displayed in both kernel and DPDK modes
using vrouter --info command.

Closes-bug: #1487358
Change-Id: Ifb63c1ea5f83e9a90fcf2f716076dba2498a6a29