Using the DHCP flag from VMI/VIF
After IRB code is introduced, the DHCP flag from VIF is no more used to
decide whether packet needs to be trapped or flooded. These flags are
encoded in L2 Bridge table entries to support BMS Dhcp support as in
case of BMS, no VIF is present in TSN's Vrouter kernel correspdonding to
BMS. But introduction of l3-only mode has introduced the requirement
back to look for this flag in VIF as there would not be any L3 bridge
entries in L3-Only mode. The fix is to check for the Dhcp enable flag in
VMI then followed by L2 bridge entry.

Alon with the above change, the code is reorganised to use the fmd's
mcast source value. In absence of fmd_src in FMD, when packet is moving
from L2 multicast nexthop to Fabric nexthop, to identify the whether
packet is received from Tor or another Vrouter, validate_src is invoked
again to identify the source of the packet, leading to redundant checks.
As part of IRB code fmd_src is introduced to identify which which source
the packet has been received and this is reused to identify the source.

closes-bug: #1486490
closes-bug: #1491051
closes-bug: #1492529

Change-Id: Id188ff821ccebbc12f42dbe4755a11c39e838317