Reinitialising rt req structure for second lookup
When the inet route lookup is done, the lookup returns the prefix length
of the route the mtrie leaf has. If the same route request is reused to
do another lookup, look up fails as prefix length is no more 32 and
lookup routine expects the prefix length to be 32 bits for ipv4.

As a fix, after the first lookup, the prefix length is reinitialised to
32 again.

Change-Id: I54622001a6110d3d69d57487281619a2641ffe92
closes-bug: #1488528