Reinitialising rt req structure for second lookup
When the inet route lookup is done, the lookup returns the prefix length
of the route the mtrie leaf has. If the same route request is reused to
do another lookup, look up fails as prefix length is no more 32 and
lookup routine expects the prefix length to be 32 bits for ipv4.

As a fix, after the first lookup, the prefix length is reinitialised to
32 again.

Change-Id: I370920fa74e4b8b4e6b23ed7768c8b0d6cca83d5
closes-bug: #1488528