Taking the correct ethernet header
As part of fix for bug 1486490, the l2 multicast control packets are
handled in a different function to process DHCP packets. To do the
bridge lookup, the source mac address needs to be taken from ethernet
haeder. But as the packet is already pulled, source mac is taken wrongly
and bridge lookup is going wrong. As a fix, the offset of ethernet
header is corrected now.

Also, as fabric composite nextghop does not have any more multicast
validate source function, this is also removed.

Change-Id: I353ccbb81629210ae1d328c9a3bf5c6ca965a8c2
closes-bug: #1491051