Dont trap Neighbour solicitation packets
When IPV6 multicast packet is deep inspected, if the packet is
identified as Neighbour discovery, this should no more be trapped. But
as the packet is identified as control packet and not DHCP the code is
ending up in Trapping this to Agent. This bug is introduced as a fix for
1486490.

Right now neighbour discovery is no more trapped, and processed
regularily.

Change-Id: Id188ff821ccebbc12f42dbe4755a11c39e838317
closes-bug: #1492529