Increasing link local port range
As of now Vrouter uses link local port range from 32768 till 65535. This
port range is default port range in Linux for ephemeral ports. This
range can be modified using sysctl. If modified using sysctl to a
different range, Vrouter still uses the old range and this leads to
failure of link local service if port allocated by Agent is out side the
above default range.

As a fix, complete port range of 0 till 65535 is used. Also new utility
"llocal" is added which dumps the existing set ports of TCP and UDP for
better debugging

Change-Id: I72a708b288cc6cb36bf3097ab87c11aebe71ca59
closes-bug: #1492408