ncreasing link local port range
As of now Vrouter uses link local port range from 32768 till 65535. This
port range is default port range in Linux for ephemeral ports. This
range can be modified using sysctl. If modified using sysctl to a
different range, Vrouter still uses the old range and this leads to
failure of link local service if port allocated by Agent is out side the
above default range.

As a fix, complete port range of 0 till 65535 is used and a debug
utility "llocal" is added to dump the ports

Change-Id: I4283b97e5665cd56cae4157e1a01366de90b04ae
closes-bug: #1492408