Limit the number of component nexthops that are dumped
If the number of component nexthops are huge, with the current logic
of allocating memory for the encoding buffer in vrouter, a dump
operation can fail because of insufficient memory. For such a huge list
it doesn't make sense to dump all the component nexthops in a DUMP
operation. We will limit the dump to a smaller number and ask users to
do a GET operation of the parent nexthop to get the complete list.

Closes BUG: 1466607

Change-Id: Ia774eba66364a8aa761d8544fee3911a37e39a79