Adapting flow table to generic hash table
Currently the over flow table of flow table is a different table and is
searched if lookup does not succed in main table. The over flow table is
searched completely starting from a hash till the entry is found. This
leads to performance drop if entry does not exists in the table as such.

As a fix, the flow table is adapted to generic hash table, which
maintains a linked list of entries from the last bucket entry of main
table. So the search always looks into only these entries rather the
complete table.

Change-Id: I494a57696c99592e4d8509071efb8b06cd48829c
closes-bug: #1497833