Check for presence of v6 header before accessing
vRouter has assumed that the presence of a v6 l4 protocol type also
means the presence of a corresponding v6 IP header. This obviously
can't be true for packets that deliberately exploit this logic. Hence,
fix that assumption by explicitly checking for the presence of v6
header before accessing it.

Change-Id: Ifad6cc912d5aa68144f9f481516dce2d1dcee5f7
Partial-BUG: 1504710