Returning key length in VIF Bridge get_key callback
While searching hash table, if a particular entry is active entry,
to compare the key with search key, has entry's get_key() is invoked.
This call back is supposed to return the key length to do a memcmp by
hashtable. Returning 0 length neglects the hash entry for comparison. In
the Vcenters VIF Bridge table, the get_key() was not returning right key
length resulting in wrong search leading to issues.

As a fix, the size of key is returned.

Change-Id: I4ae4c92a943001b23307865e1601d72913af753e
closes-bug: #1511318