Maintaining free overflow entries as linked list
As of now the over flow entries are represented using bitmap. To
allocate a free entry this bitmap is searched for free bit. This can
lead to performance issues if most of the entries are used and very few
entries are empty, as there will be search of complete bitmap.

To overcome this, the free overflow entries are maintained as a linked
list, and entry from head of the list is used to allocate one. To
release it, the entry is added back at head.

Change-Id: I20343a0be290c540a935adac4ef2614bf5f955e4
closes-bug: #1497833