Mirroring the GARP packes rather trapping
When the Agent puts the Trap flag in L3 route it expects the Garp
packets to be trapped to it. As of now, the packet is only getting
Trapped and not getting flooded when this flag is seen. As such, Garp
packet always need to be flooded irrespective of the Trap flag.

To fix this, when the Trap flag is seen in the L3 route, a copuy of Garp
packet is given to VM and original copy is flooded as well.

Change-Id: I586385e157cd221bfa21d879b7601ebfa00cd5c8
closes-bug: #1511237