DPDK: fix a deadlock in vrouter-dpdk as a result of an agent crash
The problem seems to be that the pkt0 interface was deleted, but
the packet thread was stuck in an poll().
The KNI thread was waiting on the dpdk if lock, which was held by the
netlink thread.
The netlink thread was stuck in
vr_dpdk_lcore_cmd_post_all(VR_DPDK_LCORE_TX_RM_CMD,…) because the
earlier VR_DPDK_LCORE_RX_RM_CMD was not picked up the packet thread.

Change-Id: I36cb877f666c08c5f0957a32c7b2ba81b44deb0a
Closes-bug: 1514899