Enabling ARP processing on Service interface
To avoid the flooding of packets on service interface, BUM traffic on
service interface is drooped. This is fine as service instance does not
communicate to another sevice instance. This fix enables the ARP processing on
service instance interface so that it can communicate to another service
instance. If ARP algorithm choses to flood the ARP request, it is still
dropped, so that it would not loop.

Change-Id: I31e18f16d8baa5d9dd65da202636b36048b64a0f
closes-bug: #1516914