Returnign valid hash entry in vr_get_hentry_by_index() API
Bridge and flow modules that are using the hash tables, are finding the
hash entry by index using above API. The API returns entry without
verifying whether the entry is valid or not. The moduels use the entry
if not NUUl, as if they are valid entries. This can either lead to
misbehaviour or even a crash.

As a fix, the API is modifies to return only if it is Valid entry. A new
API, __vr_get_hentry_by_index(), is introduced  to return the entry
unconditionally even if invalid.

Change-Id: I7a0c943a56da7314a718105e6a08cca064f9bfef
closes-bug: #1517749