DPDK: Fix vhost connectivity with VLANs enabled
A VLAN tag should not be passed to dp-core processing in case when the packets
are received on the fabric interface with the same VLAN as vRouter is
configured to.

We do not have a separate vlanX interface, as in the kernel vRouter case,
which strips the VLAN tags and makes vRouter receiving clean ethernet frames
(which ensures no VLAN tag is passed to dp-core for packets from fabric
interface).

We do stripping in vRouter itself, and thus have to take care of this by
removing the flag.

Change-Id: I392175784c5d9ea4f72924c2c3c6e2c07b103b3e
Partial-Bug: #1508153