Not invalidating the overflow list while deleting bucket entry
As part of fix for Bug 1520614, this bug has been introduced, where the
next pointer from main hash entry is wrongly made NULL while deleting it,
leaving the over flow list non accessible.  If this non accessible over
flow entry is also marked Deleted and if Agent issues a soft reset,
these entries are not processed any more as these are not accessible
from main bucket entry.

As a fix, the next pointer of bucket entry is not marked NULL

Change-Id: Id1cb9d885c2a2f6377aca4193dc70044d4c08c5c
closes-bug: #1522765