Set appropriate TCP flags in reverse flow at creation
When agent creates a reverse flow for an existing forward flow,
appropriate TCP flags should also be set based on the TCP flags
that are set in the forward flow. Otherwise, eviction
(and other features that depend on flags in both forward and
reverse entries) might not work. Case in point is the D flag.

Any fragment other than the first fragment of the packet should
not be allowed to create a new flow.

While creating a defer call back, unset the evict flags only if
the context that led to creation of the defer was because of
eviction. Otherwise, there could be some misbehavior.

Allow eviction for flows even if there is only one way link
between forward and the reverse flow

Change-Id: I7bccd256e4d33eaf4623a89e49c40d0928c372c4
Closes-BUG: 1362701