Reset evicted flows too at the time of soft-reset
Evicted flows, while active in the flow table, is not active in the
database used by entry allocator. The API that is used to get entries
that need to be reset returns only active entries in the allocator
database, thus excluding the evicted flows from the process of reset.
Introduce a new API that will return all entries that are active
in the flow table and use that api from the reset function.

Change-Id: I41dc242dcc308768049737c644f7257f0b581ad7
Closes-BUG: 1524622