Trapping DHCP packets in the ingress processing of Tap interface
Right the DHCP packets are processed in L2 multicast nexthop as DHCP
request packet is a multicast packet. In a scaled setup, this leads to
issues, as Agent receives the config late and VRF correspodning to VMI,
L2 broadcast route and multicast nexthop also does not get added. This
leads to DHCP request packet getting dropped.

As "D" flag is set on VMI by Agent when nova-compute plugs the
interface, the DHCP packet can be trapped to Agent without even the
configuration. Agent in this case can provide the short lease to VM.

So as a fix, the DHCP processing both in V4 and V6 are moved from L2
multicast nexthop to Vm ingress processing.

Change-Id: Ibf7e6a6b7c1201ff8c487ad561a9245831b81b51
closes-bug: #1525028