If needed, DP originated packets should undergo csum/gso updates
For packets that are tunneled, vRouter updates the checksum offset of
the packet to reflect tunnel addition. For mirrored packets, if the
mirror destination (analyzer) is local, no tunnel headers
(MPLSO(UDP/GRE)) are added and hence a checksum offset update is not
done. This behavior is wrong since we do add mirror headers (and hence a
tunnel) and the whole packet needs a checksum value.

Change-Id: Ia8c184d0980e7880bff90ff53f178e657a6be311
Closes-BUG: #1528460