Trap packets that need "ICMPv6 packet too big"
IPv4 packets that have DF bit set and needs fragmentation are trapped
to agent so that agent sends the appropriate error to the source. For,
IPv6, till now we were sending ICMPv6 responses for such packets from
DP itself. Move that functionality to agent so that agent can send
responses with appropriate addresses (for e.g.: gateway address).

Change-Id: I16cb9cd4b275fdb59eef050bc504e2b1b9766660
Partial-BUG: #1424995