Trap the packet if the RA bit is set in the vxlan header
RA or the 'Router Alert' bit will be set for packets that needs to
be analyzed by the terminating VTEP rather than the destination end
point. For diagnostic ping and traceroute, RA bit will be set and
will need to be processed by agent to analyze and respond back. Hence,
trap such packets.

Change-Id: I5ab73fed3bc2f6beb15d5afbc1cf8181d2a34cbe
Partial-BUG: #1533056