DPDK: fix KNI interface scheduling
Currently the KNI interface is assigned to core 0 as vhost0 is created
before all VM interfaces. As it has only one queue, there are rings
from other lcores to send packets to core 0. As a result of this, the
first few VM interfaces are not assigned to core 0. They are assigned
to core 1,2,3 and then 1,2,3 again.

Change-Id: I530d1dd4537a69e66be25a07e5ae5f7f31d16bab
Closes-bug: 1535329