DPDK: fix setting MTU on bonded interface.
Unnecessary setting the PMD flag on a vhost interface lead
to an issue, where with bond interface configured MTU could
be changed. Doing ifconfig vhost0 mtu XXX resulted in setting
MTU to the virtual bond device, instead of settin it to bond
slaves.

Fixing this revealed that there is no bond support in the callback
configuring the device up and down, so this is commit fixes this
as well.

Change-Id: I39097eb08726e42ad77d17595719b3d187320f39
Closes-bug: #1537190