Allow configuration of outstanding hold flow limit
If the flow creation rate exceeds the rate at which the creation
is processed by agent, then the number of hold flows in the vRouter
will keep increasing. Such a scenario, in addition to increasing
the load of the agent, thus deteriorating an already glum situation,
will also end up in the complete occupation of the flow table,
leading to a complete traffic blackout for future flows.

To prevent such an abrupt degradation, we throttle the creation of
flows such that the number of flows that are in the HOLD state does
not exceed a limit, thus reducing the load on the agent and hence
leading to faster flow processing times. Currently, that limit is not
configurable.

There is a requirement to make that limit configurable/tunable so that
a tuned value can be arrived for various deployment scenarios.

Change-Id: Ic8c7811dfe83ae41d77b96401d0edc70bc4067db
Closes-BUG: #1540218