ICMPv6 errors should match the flow that led to the error
All ICMPv6 errors should follow the flow that originally led to the
error in the first place. Thus, even if ICMP packets are not allowed,
error packets will make it to the source of the original flow,
resulting in notifications to the source and allowing it to complete
its functionality.

Change-Id: Ie0baebc9efc97367e75c798a4da2c8532c905d19
Closes-BUG: #1539724