Reverting from IANA MplsoUdp port
The MplsoUdp port is moved to IANA port (6635) while transmitting, with
the view that receiving vrouter would be able to handle both old (51234)
and new ports even across releases. But new port reception is fine
only in R3.0 and not in other releases.

So the port is moved back to old value and reception is handled well for
both old and new ports in 3.0. Transmitting port would be moved back to
new in later releases.

Change-Id: Ib9c5151702c1254df4b21c370031a19b3874b0eb
partial-bug: #1420900