Fix 1: In case of fragmentation with VLANs, there was a bug in the datapath where after inserting the VLAN header in the packet mbuf structure, the pkt datastructure was not updated using pkt_push() call. This was causing corruption to the packet header when fragmentation was happening in dpdk_fragment_packet(). Fixed the same.
Fix 2:
In case of bonding, the rx error counter was incorrectly being incremented even though
packets were correctly received by the NIC. The XEC error counter in the NIC was
incrementing due to 0 UDP checksum. In such cases, we need to subtract the XEC count
to get the acutal error count. This was not happening in bonded case. Fixed the same.

Change-Id: I9e36af57859788a6c1dabae5e5a5bfabb90a79fe
Closes-bug: #1532319