Disallow a new sub-interface for an existing vlan
When sub-interfaces with the same vlan id are added again with a
different vif index, we overwrite the existing entry in the vlan
table with the new vif. On deletion of the old vif, we check whether
the pointer in the vlan table match or not. If it does not match,
we do not remove the napi setup, but go ahead and free the napi
structure. This could potentially result in memory corruption.

To avoid such scenarios, return an error if a vif already exists
for a particular vlan

Change-Id: I12b854ab982e122eb3c2c7ce8bc32f67d4aea940
Partial-BUG: #1544832