List flows based on match criteria
Add a new option "--match" to match flows based on certain criteria.
With the match option, one will be able to match based on source ip
or destination ip (with or without ports) or vrf or protocol or any
of them together with any other of them.

e.g.:
--match 1.1.1.1:20
--match "1.1.1.1:20,2.2.2.2:22"
--match "[fe80::225:90ff:fec3:afa]:22
--match "10.204.217.10:56910 & vrf 0 & proto tcp"
--match "10.204.217.10:56910,169.254.0.3:22 & vrf 0 & proto tcp"

Change-Id: Id70e6d60950babe44612e6aded036c5d1bec6f92
Partial-BUG: #1513001