DPDK: fix (again) hashing of fragmented packets
We can not hash L4 ports as fragments other than the first have
no L4 headers. In result fragments of the same packets may have
been distributed onto different lcores and potentially be routed
out of order.

The fix was originally committed by Michal Dubiel (commit d59ca668)
For a reason the fix was dropped in the commit 1056396f

Change-Id: Iac607e6e14a139a0c015eeb5c9ff8c3de04854c6
Closes-Bug: #1474432