'GET' option for 'rt' command
A '--get' switch for the 'rt' to match a particular route and dump it.
Currently, only for inet and inet6 routes.

Change-Id: I3b53dda635f4f4af5b7f79ddf47e14313f372f81
Partial-BUG: #1548757

Support GET for L2 routes

Support 'GET' operation for a specific route in the vRouter bridge
table too.

Also, support giving a specific prefix length to the route and return
a match based on that specific prefix.

Change-Id: I93980925e70f45482c372ba3beeca8bfbd3028b2
Closes-BUG: #1548757