Dont mirror the same packet again
If mirroring is enabled on analyzing VM's port itself, one copy of the
original packet gets mirrored to that port. But this mirrored packet
should not be mirroed again. Right now without verifying whether the
packet is mirrored or not, it is attempted for mirroring again.

When a packet is mirrored it is marked with flag VP_FLAG_FROM_DP. If
this flag is set it is not mirrored again.

Change-Id: Ie4cca916945878b77bc4db7666992f3933db3e13
closes-bug: #1549761