Keeping ingres vif id in L2 header for post GRO processing
Currently the L2 header used for both L2 and L3 GRO processing is only
next hop id. Post GRO, the nexthop is retrieved from this id and packet
is subjected to this nexthop for further processing. The ingress
interface filled in the packet is nh_dev of this nexthop. This is
resulting in ingress interface as Tap interface though packet acttually ingressed
on Fabric interface. This can potentially lead to issues in the packet's
further processing as Tap interface is not the real ingress
interface.

As a fix, the ingress interface id is also pushed as L2 header.

Change-Id: I6fbe055fff7da4c957fde29d132b56ea19bd2f3f
closes-bug: #1550632