Waiting for completion of asynchronous events as part of soft reset
Right now when Agent comes up, it does Vrouter soft reset. As part of
soft reset, voruter stops the traffic, reinitialises all the data
structures and starts processing the traffic again. Before it starts the
traffic, it needs to wait for the completion of already scheduled
asynchronous events like RCU callbacks and WaitQueue events. Without
which, the asynchronous call backs might act on the old data, leading to
issues.

As a fix, as part of soft reset call from agent, Vrouter waits for
asynchronous events to complete before start processing the traffic
again.

Change-Id: I33ac32d20798a5bc07858e7e8d66b9ffb60cd775
closes-bug: #1561346